requirements.txt
discord.py >= 0.16.8
requests>=0.11.1